x86 UEFI: Increase default QEMU memory size
QEMU default memory size is 128MiB. After UEFI boot up and we call
`LibMemoryMap()`, the largest consecutive memory region is ~42MiB.
Although this is sufficient for many test cases to run in UEFI, some
test cases blindly allocate more memory and triggers the following
assert failure:

   lib/vmalloc.c:153: assert failed: pa

This is because the memory allocator runs out of allocatable memory and
cannot fulfill the allocation request. For e.g., `x86/pmu.c` tries to
blindly allocate ~60MiB memory that exceeds the ~42MiB allocatable
memory. But in Seabios, the allocatable memory is ~120 MiB and is
sufficient for ~60MiB allocation.

A simple fix is to increase the QEMU default memory size to 256MiB, so
that the largest allocatable memory region in UEFI is large enough.

After this commit, test case `x86/pmu.c` can run in UEFI and generate
the same output as in Seabios.

Change-Id: I20e8acc47706ede08a5a8d77891a9927efbaae72