x86 UEFI: Print runtime base address on error
UEFI loads EFI application binaries to dynamic runtime addresses, so
given a runtime address, it's hard to find out the corresponding address
in disassembly code.

This commit provides code to print out runtime base address when
unexpected exception happens, so that we can calculate the offset
address and find it in disassembly code.

Change-Id: Ic30de85f4a19483ebc06174dc98633f679633d53