x86 AMD SEV-ES: Check SEV-ES status
This commit provides initial start up code for KVM-Unit-Tests to run in
an SEV-ES guest VM. This start up code checks if SEV-ES feature is
supported and enabled for the guest.

In this commit, KVM-Unit-Tests can pass the SEV-ES check and enter
setup_efi() function, but crashes in setup_gdt_tss(), which will be
fixed by follow-up commits.

Change-Id: Ibbf471e6a56ba57e1f0edd76aa76b9b51c5e9647
Signed-off-by: Zixuan Wang <zixuanwang@google.com>