x86 AMD SEV-ES: Set up #VC without UEFI segments
In the current SEV-ES set up code, KVM-Unit-Tests copies the UEFI code
and data segments from UEFI GDT into KVM-Unit-Tests GDT. This was to
make sure the UEFI #VC handler has the correct segments to use.

But it is possible to modify the #VC handler IDT entry and update its
'selector' field which stores the code segment for the handler. In this
way, UEFI #VC handler can use KVM-Unit-Tests code segment, and does not
require UEFI segments to be copied into KVM-Unit-Tests GDT.

Change-Id: I4a783401e8ddf998e24dc5dc3936d716ddd3ccd2